Completely overwrite the settings file on save
File.OpenWrite() does not truncate the file, so if a smaller amount
of data was written (i.e. using a smaller channel name), the data not
overwritten was left at the end of the file.

Using File.Open() with FileMode.Create will ensure that if the file
exists, it will truncate the file.